requirements.txt
python-binance==1.0.19
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
aiohttp==3.9.1
asyncio==3.4.3
